# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: API CineVault
  description: Documentación oficial de la API de CineVault (red social de cine).
  version: 1.0.0
  contact:
    name: Equipo CineVault
servers:
  - url: https://api.cinevault.art/api
    description: Producción
  - url: http://localhost:4000/api
    description: Desarrollo local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
  schemas:
    UsuarioPublico:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [admin, editor, user] }
        avatar_url: { type: string, format: uri, nullable: true }
        bio: { type: string, maxLength: 300, nullable: true }
        is_verified: { type: boolean }
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
        code: { type: string }
...
  responses:
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /users:
    get:
      tags: [Usuarios]
      summary: Listar usuarios
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Lista de usuarios, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/UsuarioPublico' } } } } }
  /users/{id}:
    get:
      tags: [Usuarios]
      summary: Perfil público
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Perfil público }
        '404': { $ref: '#/components/responses/NotFound' }
  /users/follow/{id}:
    post:
      tags: [Usuarios]
      summary: Seguir usuario
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Seguido }
  /users/unfollow/{id}:
    delete:
      tags: [Usuarios]
      summary: Dejar de seguir
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Unfollow ok }
  /users/{id}/followers:
    get:
      tags: [Usuarios]
      summary: Seguidores
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Lista de seguidores }
  /users/{id}/following:
    get:
      tags: [Usuarios]
      summary: Siguiendo
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Lista de seguidos }
...
  /favorites/user/{userId}:
    get:
      tags: [Favoritos]
      summary: Favoritos de un usuario
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Lista }
...
  /reviews/user/{userId}:
    get:
      tags: [Reseñas]
      summary: Reseñas de un usuario
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Lista }
...